+++ NEWS.md
@@ -4,7 +4,8 @@
 * Use `censusapi` instead of `tidycensus` for fewer dependencies
 * New `estimate` vector type that tracks uncertainty through mathematical operations 
 * Expose parsed tables to users with a new `cens_table` type
-* Improve tidiers, including income bin tidiers
+* More flexible geography options when downloading data. See `cens_geo()` for details.
+* Improved tidiers, including new income bin and detailed race tidiers
 
 # easycensus 0.2.0